ASoC: apple: Add macaudio machine driver
This commit adds an ASoC machine driver for Apple Silicon Macs.

Each supported Apple Silicon Mac implements some combination of:

    - Apple's I2S peripheral (MCA)
    - Apple-specific variants of TI TAS2770 *or* TAS2764
    - An Apple-specific variant of Cirrus CS42L42 (CS42L84)

The machine driver takes an I2S cluster from MCA and combines it
with some subset of the codecs present on the machine. The combinatorics
are machine-specific and expressed as DAI links in the devicetree.

Most Apple Silicon machines with TAS2764 also allow the codecs to
send their V/ISENSE data back to the SoC via a second I2S cluster. This
is implemented as a capture PCM so that a userspace daemon can
implement a variant of TI's Smart Amp speaker protection algorithm
in software. Since these machines have no hardware failsafe preventing
speaker damage, the machine driver will mute the speakers and lock
the volume kcontrols if the userspace daemon is not running.

Co-developed-by: Hector Martin <marcan@marcan.st>
Co-developed-by: James Calligeros <jcalligeros99@gmail.com>
Co-developed-by: Janne Grunau <j@jannau.net>
Signed-off-by: Martin Povišer <povik+lin@cutebit.org>
Signed-off-by: Hector Martin <marcan@marcan.st>
Signed-off-by: Janne Grunau <j@jannau.net>
Signed-off-by: James Calligeros <jcalligeros99@gmail.com>