ASoC: apple: Add macaudio machine driver
Add a machine driver for Apple Silicon Macs.

This driver includes a novel mechanism for guaranteeing speaker safety.
the TI TAS2764 and TAS2770 codec variants used in Apple Silicon Macs do not
implement any hardware safety features, beyond feeding V/ISENSE data back.
It is therefore possible to permanently damage the speakers found in these
devices, particularly the tweeters.

Apple deals with this by having CoreAudio implement speaker safety in userspace.
The V/ISENSE data is fed back to CoreAudio, which then uses the Thiele/Small
parameter set for each speaker to model its voice coil and magnet temperature.
If these temperatures approach unsafe levels, CoreAudio transparently limits the
volume until they cool down. This allows Apple to maximise headroom during normal
use, as they do not have to hard limit the volume to a known-safe level.

Thankfully, the Thiele/Small parameters for every single Mac are stored in plain
text .plist files in the macOS root filesystem. Since these parameters are physical
properties of the hardware, they are not copyrightable and we can simply take them
and use them to develop our own speaker safety system. We do this by locking the
volume kcontrol for each speaker, and giving exclusive control to a userspace daemon.

A secondary PCM feeds the raw V/ISENSE data back to userspace. Our safety daemon
opens this PCM and reads the data. The driver has a failsafe interlock - the daemon
must write a magic value to a kcontrol in order for the driver to unlock the
speaker volume kcontrols. In their locked state, the volumes are limited to a very
low level so that we can still guarantee runtime safety in the event that the
daemon fails.

Co-developed-by: Hector Martin <marcan@marcan.st>
Signed-off-by: Hector Martin <marcan@marcan.st>
Co-developed-by: Janne Grunau <j@jannau.net>
Signed-off-by: Janne Grunau <j@jannau.net>
Co-developed-by: James Calligeros <jcalligeros99@gmail.com>
Co-developed-by: Sasha Finkelstein <fnkl.kernel@gmail.com>
Signed-off-by: Sasha Finkelstein <fnkl.kernel@gmail.com>
Signed-off-by: Martin Povišer <povik+lin@cutebit.org>
Signed-off-by: James Calligeros <jcalligeros99@gmail.com>