ASoC: macaudio: constrain frontends with multiple backend codecs
In order to support a wide variety of machines, it is necessary
for the frontends to have enough TDM slots to supply the backend
codecs they are routed to. When a FE has more TDM slots than its
backend has valid codecs, userspace opens too many channels,
and frames are lost to the aether.

Dynamically constrain the FEs when they are routed based on the
number of codecs they are supposed to be linked to, as laid out
in the devicetree. Note that this is only relevant for the internal
speaker array, and does not affect the headphone jack.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>