ASoC: macaudio: constrain frontend channel counts
In order to support a wide variety of machines, it is necessary
for the frontends to have enough TDM slots to supply the backend
codecs they are routed to. This causes problems on machines with
less backend codecs than frontend TDM slots - userspace will try
to open channels on all slots, resulting in lost frames and garbled
sound.

We must therefore dynamically constrain the frontends at runtime
such that userspace can never open more channels than are actually
present on the hardware represented by a given frontend PCM.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>