OPP: decouple properties in opp_parse_supplies()
Allow platforms to specify and consume any subset of opp-microvolt,
opp-microamp, or opp-microwatt without depending on having opp-microvolt
specified in each OPP.

This is useful on some platforms, such as Apple Silicon, which do not
expose their voltage regulators beyond firmware but can take advantage
of opp-microwatt to provide Energy Aware Scheduling. This was an implicit
use case of opp-microwatt, making the inability to do so prior a bug
in its implementation.

Fixes: 4f9a7a1dc2a2 ("OPP: Add "opp-microwatt" supporting code")

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>