hwmon: applesmc-arm: handle invalid keys in devicetree
the SMC firmware interface is unstable. Apple may remove keys from
time to time. If these invalid keys are present in the device tree
and the driver tries to reference them, we will get a NULL pointer
error.

Gracefully handle invalid keys by continuing through the loop even
if a broken key is found, and setting the number of *valid* keys
after we have parsed all values in the devicetree node.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>