drm: apple: respect drm_plane_state zpos
The for_each_oldnew_plane_in_state iterator is nondeterministic
in terms of the order of planes. DCP expects surfaces to be
fed to it in the correct order. Relying on the iterator to
lazily increment the index into surf[] means we cannot meet
this expectation. The constant reordering of planes in
the surf[] array seems to cause DCP to crash under certain
circumstances. Cursors will also often be drawn under the main
plane, which is less than ideal.

By fortuitous circumstance, DCP's z-ordering matches DRM's, which
is what userspace also expects!

Populate surf[] in the order everyone expects us to. This fixes
a whole host of odd behaviour when wiring up multiple DRM universal
planes.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>