arm64: apple: rework hwmon dtsi hierarchy
Keys that were assumed to be common to all machines are, in actuality, not. This means
that we cannot avoid having SKU-specific keys at the .dts level.

Rearrange how the .dsti files collect each other so that we don't need to include
anything at the device level. This makes it slightly cleaner to add device-specific
keys.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>