ASoC: cs42l84: delay headset detect to prevent misdetections
The tip sense interrupt is quite trigger happy. There are instances
where we kick off - and finish - headset detection before the user
has had a chance to even finish plugging in the jack. This leads to
misdetections, including detecting headsets as not having a microphone.
The only way to fix this is to unplug the headset and plug it back in.

Give the user 1.5 seconds to fully plug in their headset/headphones
before running the headset/button detection logic. This reduces the chance
of misdetections.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>