drm: apple: move plane operations into plane.h and plane.c
The DCP driver does not follow KMS atomic best practices. This was
fine when we were treating the entire display as basically a dumb
framebuffer, however this approach is becoming cumbersome and
a hinderance to performance now that we are starting to make better
use of it.

DCP's surfaces (planes in KMS parlance) are capable of a lot more than
just dumping out ARGB framebuffers, so in preparation for implementing
some of these features, pull out the current skeleton plane operations
into their own area.

Co-developed-by: Janne Grunau <j@jannau.net>
Signed-off-by: Janne Grunau <j@jannau.net>
Signed-off-by: James Calligeros <jcalligeros99@gmail.com>