drm: apple: limit transfer function and colourspace selection to YUV
RGB surfaces do not expose a colour encoding, leaving the plane's colour
encoding property set to 0. This means we were always selecting the BT.601
primaries, which was wrong.

Default to NATIVE and SDR for colourspace and transfer function
respectively when the surface is RGB. This will be expanded upon once HDR
surfaces are being blended in hardware.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>